requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0
pyyaml>=6.0
tensorboard>=2.14.0
comet-ml>=3.35.0
pytorch-metric-learning>=2.3.0

# AST parsing
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-java>=0.20.0
tree-sitter-cpp>=0.20.0
tree-sitter-javascript>=0.20.0
tree-sitter-go>=0.20.0
tree-sitter-c-sharp>=0.20.0

# Additional utilities
scipy>=1.11.0
seaborn>=0.12.0
matplotlib>=3.7.0
plotly>=5.17.0
pyarrow>=13.0.0
fastparquet>=2023.8.0

# Code analysis
radon>=6.0.1
lizard>=1.17.10
pylint>=3.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.9.0
flake8>=6.1.0
ipython>=8.15.0
jupyter>=1.0.0
